+++ Pigeonhole/minimizers.h
@@ -9,7 +9,7 @@
 
 extern vector<pair<string, int> > alphabetRef;
 
-unsigned long int extractRanking(string kmer);
+unsigned long long int extractRanking(string kmer);
 uint64_t inthash_64(uint64_t key, uint64_t mask);
 unsigned int getMinimizerRank(string windowSeed, int q, int windowSize);
 //multimap<vector<int>, int> generateMinimizers(string stringDNA, int q, int w, int m);
