+++ Pigeonhole/pigeonhole.h
@@ -20,6 +20,6 @@ extern int numberReads;
 extern int numberAcceptedReads;
 
 string reverseComplement(string read);
-void verification(vector<unsigned long int>& forwardFound, vector<unsigned long int>& reverseFound, vector<unsigned long int>& exactFound);
+void verification(vector<unsigned long long int>& forwardFound, vector<unsigned long long int>& reverseFound, vector<unsigned long long int>& exactFound);
 
 #endif //MP_PIGEONHOLE_PIGEONHOLE_H
